Address review: extract stage_events.py to break import cycle
Pulls STAGE_* constants, stage_timer, and emit_stage_event out of
provisioning.py into a new stage_events module. Both provisioning.py
and provisioning_stages.py now depend on stage_events; the deferred
bottom-of-file import (and its noqa) is gone.

Also adds __all__ to provisioning_stages.py and a comment on the
provisioning.py re-export block clarifying that the stage names are
re-exported there so tests can patch e.g. provisioning.install_runtime
to inject failures into provision_session.
