+++ src/agent_on_demand/session_service/provisioning.py
@@ -10,10 +10,7 @@ expensive shell work into a single round trip.
 
 from __future__ import annotations
 
-import contextlib
 import logging
-import time
-from typing import Iterator
 
 from sprites import Sprite, SpriteError
 
@@ -21,75 +18,10 @@ from agent_on_demand.observability import get_tracer
 
 from .client import best_effort_delete, require_client
 from .errors import ProvisionError
-from .specs import SessionSpec
-
-logger = logging.getLogger(__name__)
-
-# Stage names emitted both as `ProvisionError.stage` tags (server-side logging)
-# and as `stage` SSE events (see site/docs/api/streaming.md). Keep in sync.
-STAGE_CREATE_SPRITE = "create_sprite"
-STAGE_INSTALL_RUNTIME = "install_runtime"
-STAGE_NETWORK_POLICY = "network_policy"
-STAGE_ENV_FILE = "env_file"
-STAGE_GIT_CREDENTIALS = "git_credentials"
-STAGE_PROVISION_SETUP = "provision_setup"
-STAGE_RUNTIME_CONFIG = "runtime_config"
-STAGE_SKILLS = "skills"
-STAGE_RUNTIME_START = "runtime_start"
-
-
-def emit_stage_event(
-    session_id: str | None,
-    stage: str,
-    state: str,
-    duration_ms: int | None = None,
-    message: str = "",
-) -> None:
-    """Write a stage row to AgentSessionLog. No-op if session_id is None (unit
-    tests that exercise provision_session directly without a real session row)."""
-    if session_id is None:
-        return
-    from agent_on_demand.models import AgentSessionLog
-
-    AgentSessionLog.objects.create(
-        session_id=session_id,
-        kind="stage",
-        stage=stage,
-        state=state,
-        duration_ms=duration_ms,
-        data=message,
-    )
-
-
-@contextlib.contextmanager
-def stage_timer(session_id: str | None, stage: str) -> Iterator[None]:
-    """Emit `started` entering and `done` on clean exit, or `failed` (with the
-    exception message) on error. `duration_ms` is attached to done/failed."""
-    emit_stage_event(session_id, stage, "started")
-    start = time.monotonic()
-    try:
-        yield
-    except Exception as e:
-        emit_stage_event(
-            session_id,
-            stage,
-            "failed",
-            int((time.monotonic() - start) * 1000),
-            message=str(e),
-        )
-        raise
-    else:
-        emit_stage_event(
-            session_id,
-            stage,
-            "done",
-            int((time.monotonic() - start) * 1000),
-        )
-
-
-# Imported after STAGE_* / stage_timer are defined because provisioning_stages
-# imports those names back from this module.
-from .provisioning_stages import (  # noqa: E402
+# Re-exported so call-sites in `provision_session` resolve through this
+# module's namespace — that lets tests patch e.g. `provisioning.install_runtime`
+# to inject failures without reaching into `provisioning_stages` directly.
+from .provisioning_stages import (
     apply_network_policy,
     install_runtime,
     run_provision_setup,
@@ -98,6 +30,10 @@ from .provisioning_stages import (  # noqa: E402
     write_runtime_config,
     write_skills,
 )
+from .specs import SessionSpec
+from .stage_events import STAGE_CREATE_SPRITE, stage_timer
+
+logger = logging.getLogger(__name__)
 
 
 def provision_session(user, spec: SessionSpec, session_id: str | None = None) -> Sprite:
