+++ src/agent_on_demand/session_service/tasks.py
@@ -47,14 +47,9 @@ from agent_on_demand.models import (
 from agent_on_demand.observability import get_tracer
 
 from .errors import NoSpritesKeyError, ProvisionError, SessionHandleNotFound
-from .provisioning import (
-    STAGE_RUNTIME_START,
-    destroy_session,
-    emit_stage_event,
-    provision_session,
-    resume_session,
-)
+from .provisioning import destroy_session, provision_session, resume_session
 from .spec_factory import build_spec_for_session
+from .stage_events import STAGE_RUNTIME_START, emit_stage_event
 from .turn_argv import build_turn_argv
 
 logger = logging.getLogger(__name__)
